fix(getJWT): use controller-runtime client
